Move REAPI action cache to reapi.rs
