+++ src/rust/engine/process_execution/remote/src/remote_cache.rs
@@ -1,7 +1,6 @@
 // Copyright 2022 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 use std::collections::{BTreeMap, HashSet};
-use std::convert::TryInto;
 use std::fmt::{self, Debug};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
@@ -10,27 +9,24 @@ use async_trait::async_trait;
 use fs::{directory, DigestTrie, RelativePath, SymlinkBehavior};
 use futures::future::{BoxFuture, TryFutureExt};
 use futures::FutureExt;
-use grpc_util::retry::{retry_call, status_is_retryable};
-use grpc_util::{headers_to_http_header_map, layered_service, status_to_str, LayeredService};
 use hashing::Digest;
 use parking_lot::Mutex;
 use protos::gen::build::bazel::remote::execution::v2 as remexec;
 use protos::require_digest;
-use remexec::action_cache_client::ActionCacheClient;
 use remexec::{ActionResult, Command, Tree};
 use store::{Store, StoreError};
 use workunit_store::{
   in_workunit, Level, Metric, ObservationMetric, RunningWorkunit, WorkunitMetadata,
 };
 
-use crate::remote::apply_headers;
 use process_execution::{
   check_cache_content, populate_fallible_execution_result, CacheContentBehavior, Context,
   FallibleProcessResultWithPlatform, Process, ProcessCacheScope, ProcessError,
   ProcessExecutionEnvironment, ProcessResultSource,
 };
 use process_execution::{make_execute_request, EntireExecuteRequest};
-use tonic::{Code, Request};
+
+mod reapi;
 
 #[derive(Clone, Copy, Debug, strum_macros::EnumString)]
 #[strum(serialize_all = "snake_case")]
@@ -73,7 +69,7 @@ pub struct CommandRunner {
   append_only_caches_base_path: Option<String>,
   executor: task_executor::Executor,
   store: Store,
-  action_cache_client: Arc<ActionCacheClient<LayeredService>>,
+  provider: Arc<dyn ActionCacheProvider>,
   cache_read: bool,
   cache_write: bool,
   cache_content_behavior: CacheContentBehavior,
@@ -91,7 +87,7 @@ impl CommandRunner {
     store: Store,
     action_cache_address: &str,
     root_ca_certs: Option<Vec<u8>>,
-    mut headers: BTreeMap<String, String>,
+    headers: BTreeMap<String, String>,
     cache_read: bool,
     cache_write: bool,
     warnings_behavior: RemoteCacheWarningsBehavior,
@@ -100,25 +96,14 @@ impl CommandRunner {
     rpc_timeout: Duration,
     append_only_caches_base_path: Option<String>,
   ) -> Result<Self, String> {
-    let tls_client_config = if action_cache_address.starts_with("https://") {
-      Some(grpc_util::tls::Config::new_without_mtls(root_ca_certs).try_into()?)
-    } else {
-      None
-    };
-
-    let endpoint = grpc_util::create_endpoint(
+    let provider = Arc::new(reapi::Provider::new(
+      instance_name.clone(),
       action_cache_address,
-      tls_client_config.as_ref(),
-      &mut headers,
-    )?;
-    let http_headers = headers_to_http_header_map(&headers)?;
-    let channel = layered_service(
-      tonic::transport::Channel::balance_list(vec![endpoint].into_iter()),
+      root_ca_certs,
+      headers,
       concurrency_limit,
-      http_headers,
-      Some((rpc_timeout, Metric::RemoteCacheRequestTimeouts)),
-    );
-    let action_cache_client = Arc::new(ActionCacheClient::new(channel));
+      rpc_timeout,
+    )?);
 
     Ok(CommandRunner {
       inner,
@@ -127,7 +112,7 @@ impl CommandRunner {
       append_only_caches_base_path,
       executor,
       store,
-      action_cache_client,
+      provider,
       cache_read,
       cache_write,
       cache_content_behavior,
@@ -300,7 +285,7 @@ impl CommandRunner {
         self.instance_name.clone(),
         request.execution_environment.clone(),
         &context,
-        self.action_cache_client.clone(),
+        self.provider.clone(),
         self.store.clone(),
         self.cache_content_behavior,
       )
@@ -401,7 +386,7 @@ impl CommandRunner {
   async fn update_action_cache(
     &self,
     result: &FallibleProcessResultWithPlatform,
-    instance_name: Option<String>,
+    _instance_name: Option<String>,
     command: &Command,
     action_digest: Digest,
     command_digest: Digest,
@@ -422,28 +407,10 @@ impl CommandRunner {
       .ensure_remote_has_recursive(digests_for_action_result)
       .await?;
 
-    let client = self.action_cache_client.as_ref().clone();
-    retry_call(
-      client,
-      move |mut client| {
-        let update_action_cache_request = remexec::UpdateActionResultRequest {
-          instance_name: instance_name.clone().unwrap_or_else(|| "".to_owned()),
-          action_digest: Some(action_digest.into()),
-          action_result: Some(action_result.clone()),
-          ..remexec::UpdateActionResultRequest::default()
-        };
-
-        async move {
-          client
-            .update_action_result(update_action_cache_request)
-            .await
-        }
-      },
-      status_is_retryable,
-    )
-    .await
-    .map_err(status_to_str)?;
-
+    self
+      .provider
+      .update_action_result(action_digest, action_result)
+      .await?;
     Ok(())
   }
 
@@ -591,10 +558,10 @@ impl process_execution::CommandRunner for CommandRunner {
 async fn check_action_cache(
   action_digest: Digest,
   command_description: &str,
-  instance_name: Option<String>,
+  _instance_name: Option<String>,
   environment: ProcessExecutionEnvironment,
   context: &Context,
-  action_cache_client: Arc<ActionCacheClient<LayeredService>>,
+  provider: Arc<dyn ActionCacheProvider>,
   store: Store,
   cache_content_behavior: CacheContentBehavior,
 ) -> Result<Option<FallibleProcessResultWithPlatform>, ProcessError> {
@@ -606,7 +573,8 @@ async fn check_action_cache(
       workunit.increment_counter(Metric::RemoteCacheRequests, 1);
 
       let start = Instant::now();
-      let response = get_action_result(action_digest, instance_name, context, action_cache_client)
+      let response = provider
+        .get_action_result(action_digest, context)
         .and_then(|action_result| async move {
           let Some(action_result) = action_result else { return Ok(None) };
 
@@ -650,33 +618,4 @@ async fn check_action_cache(
     }
   )
   .await
-}
-
-async fn get_action_result(
-  action_digest: Digest,
-  instance_name: Option<String>,
-  context: &Context,
-  action_cache_client: Arc<ActionCacheClient<LayeredService>>,
-) -> Result<Option<ActionResult>, String> {
-  let client = action_cache_client.as_ref().clone();
-  let response = retry_call(
-    client,
-    move |mut client| {
-      let request = remexec::GetActionResultRequest {
-        action_digest: Some(action_digest.into()),
-        instance_name: instance_name.clone().unwrap_or_default(),
-        ..remexec::GetActionResultRequest::default()
-      };
-      let request = apply_headers(Request::new(request), &context.build_id);
-      async move { client.get_action_result(request).await }
-    },
-    status_is_retryable,
-  )
-  .await;
-
-  match response {
-    Ok(response) => Ok(Some(response.into_inner())),
-    Err(status) if status.code() == Code::NotFound => Ok(None),
-    Err(status) => Err(status_to_str(status)),
-  }
 }
